@@ -1,5 +1,6 @@
 {
   "name": "devsearch",
+  "type": "module",
   "displayName": "DevSearch",
   "description": "dev assistant for tasks beyond coding",
   "version": "0.0.1",
@@ -44,21 +45,31 @@
     "lint:fix": "eslint . --fix",
     "format": "prettier --write .",
     "pretest": "npm run lint",
-    "test": "vscode-test"
+    "test": "mocha -r ts-node/register 'test/**/*.test.ts'"
   },
   "devDependencies": {
+    "@types/chai": "^4.3.19",
+    "@types/chai-as-promised": "^8.0.0",
+    "@types/jest": "^29.5.13",
+    "@types/mocha": "^10.0.8",
+    "@types/node": "^20.16.5",
+    "@types/sinon": "^17.0.3",
     "@types/vscode": "^1.93.0",
-    "@types/mocha": "^10.0.7",
-    "@types/node": "20.x",
     "@typescript-eslint/eslint-plugin": "^6.0.0",
     "@typescript-eslint/parser": "^6.0.0",
+    "@vscode/test-cli": "^0.0.10",
+    "@vscode/test-electron": "^2.4.1",
+    "chai": "^5.1.1",
+    "chai-as-promised": "^8.0.0",
     "eslint": "^8.0.0",
     "eslint-config-prettier": "^9.1.0",
     "eslint-plugin-prettier": "^5.1.3",
+    "mocha": "^10.7.3",
     "prettier": "^3.2.5",
-    "@vscode/test-cli": "^0.0.10",
-    "@vscode/test-electron": "^2.4.1",
-    "typescript": "^5.4.5"
+    "sinon": "^19.0.2",
+    "ts-node": "^10.9.2",
+    "tsconfig-paths": "^4.2.0",
+    "typescript": "^5.6.2"
   },
   "dependencies": {
     "axios": "^1.7.7",
